Rimozione Activity intermedia per l'aggiunta di un luogo...

+++ app/src/main/java/com/example/david/ermes/Model/models/Location.java
@@ -41,7 +41,7 @@ public class Location implements Serializable{
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
@@ -49,7 +49,7 @@ public class Location implements Serializable{
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
